requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
channels==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.2
dash==2.9.2
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
Django==3.2
django-crontab==0.7.1
django-plotly-dash==2.1.4
dpd-components==0.1.0
executing==1.2.0
Faker==17.3.0
filelock==3.9.0
Flask==2.2.3
ipykernel==6.21.2
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kombu==5.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
paho-mqtt==1.6.1
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.0.0
plotly==5.14.1
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.2.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
vine==5.0.0
virtualenv==20.20.0
wcwidth==0.2.6
Werkzeug==2.2.3
